fix(gulp): minimize issues w/referencing SK code
Because all source files are required by the gulpfile during the build
stage in order to auto-generate documentation, code cannot exist at the
top execution level which assumes that SurfingKeys- or browser-related
code is defined.

Two steps were taken to minimize the issues caused by this:

1. The source files are now only included during documentation
generation, and not during normal builds.

2. Any occurrences of references to SK-specific code have been removed
from the top level of the source files.

There still remains the issue of selectively accessing SK code from the
top level of source files, e.g. as a user wishes to do in #14, but as
long as the ability to generate documentation is not desired, this
should be a lesser issue. The workaround remains to wrap SK-dependent
code in a block which is only executed in a real browser and not gulp.
[TODO: Perhaps there is a simple way to detect this, and/or we can
provide a helper function for doing so]

The Hints.js stub file is no longer necessary after this change, and so
it has been removed.